+++ CHANGELOG.md
@@ -7,13 +7,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 When publishing a new release, copy the relevant section on the [Github release page](https://github.com/Mobidec/ckanapi_harvesters/releases).
 
 
-## [Unreleased] - 2026-04-15
+## [Unreleased] - 2026-04-06
+
+
+## [0.0.26] - 2026-04-03
 
 ### Added
 
 - HTTP status code messages.
 - Disconnect at second request retry.
 - Reviewed SSL parameters for MongoDB/PostgreSQL connections.
+
+### Changed
+
+- Refactored MongoDB harvesters module names.
+
 
 ## [0.0.25] - 2026-04-01